mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "gradio>=5.22.0",
    "mcp>=1.3.0",
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
]
